move binaries and modify setup to make picca_* scripts importable
