+++ picca.egg-info/PKG-INFO
@@ -1,0 +1,10 @@
+Metadata-Version: 1.0
+Name: picca
+Version: 4.0
+Summary: Package for Igm Cosmological-Correlations Analyses
+Home-page: https://github.com/igmhub/picca
+Author: Nicolas Busca, Helion du Mas des Bourboux et al
+Author-email: nbusca@lpnhe.in2p3.fr
+License: UNKNOWN
+Description: UNKNOWN
+Platform: UNKNOWN
