+++ picca.egg-info/SOURCES.txt
@@ -1,0 +1,69 @@
+README.md
+setup.py
+bin/__init__.py
+bin/picca_Pk1D.py
+bin/picca_cf.py
+bin/picca_cf1d.py
+bin/picca_cf_angl.py
+bin/picca_co.py
+bin/picca_compute_fvoigt.py
+bin/picca_compute_pk_pksb.py
+bin/picca_deltas.py
+bin/picca_dmat.py
+bin/picca_export.py
+bin/picca_export_co.py
+bin/picca_export_coadd_zint.py
+bin/picca_export_cross_covariance.py
+bin/picca_fit.py
+bin/picca_fitter2.py
+bin/picca_fitter2_control.py
+bin/picca_fitter2_control_mpi.py
+bin/picca_mcmc.py
+bin/picca_metal_dmat.py
+bin/picca_metal_xdmat.py
+bin/picca_nersc_submit.py
+bin/picca_pk2fits.py
+bin/picca_plot_pk1d.py
+bin/picca_reduce_spall.py
+bin/picca_wick.py
+bin/picca_xcf.py
+bin/picca_xcf1d.py
+bin/picca_xcf_angl.py
+bin/picca_xdmat.py
+bin/picca_xwick.py
+picca.egg-info/PKG-INFO
+picca.egg-info/SOURCES.txt
+picca.egg-info/dependency_links.txt
+picca.egg-info/requires.txt
+picca.egg-info/top_level.txt
+py/picca/__init__.py
+py/picca/_version.py
+py/picca/cf.py
+py/picca/co.py
+py/picca/constants.py
+py/picca/converters.py
+py/picca/data.py
+py/picca/dla.py
+py/picca/io.py
+py/picca/pk1d.py
+py/picca/prep_del.py
+py/picca/prep_pk1d.py
+py/picca/utils.py
+py/picca/wedgize.py
+py/picca/xcf.py
+py/picca/fitter2/__init__.py
+py/picca/fitter2/chi2.py
+py/picca/fitter2/control.py
+py/picca/fitter2/control_mpi.py
+py/picca/fitter2/data.py
+py/picca/fitter2/effective-bins.py
+py/picca/fitter2/myGamma.py
+py/picca/fitter2/parser.py
+py/picca/fitter2/pk.py
+py/picca/fitter2/priors.py
+py/picca/fitter2/sampler.py
+py/picca/fitter2/utils.py
+py/picca/fitter2/xi.py
+py/picca/fitter2/models/DR9LyaMocks/DR9LyaMocks.fits
+py/picca/fitter2/models/PlanckDR12/PlanckDR12.fits
+py/picca/fitter2/models/PlanckDR16/PlanckDR16.fits
