+++ picca.egg-info/requires.txt
@@ -1,0 +1,10 @@
+numpy
+scipy
+iminuit
+healpy
+fitsio
+llvmlite
+numba
+h5py
+future
+setuptools
