+++ py/picca/test/test_1_deltas.py
@@ -9,7 +9,7 @@ import shutil
 import glob
 from pkg_resources import resource_filename
 import sys
-import picca_deltas
+import picca_bin.picca_deltas as picca_deltas
 
 from picca.utils import userprint
