+++ py/picca_bin/picca_reduce_spall.py
