+++ setup.py
@@ -4,7 +4,7 @@ import glob
 
 from setuptools import setup
 
-scripts = glob.glob('bin/*')
+scripts = glob.glob('py/picca_bin/picca*.py')
 
 description = "Package for Igm Cosmological-Correlations Analyses"
 
@@ -15,7 +15,7 @@ setup(name="picca",
     url="https://github.com/igmhub/picca",
     author="Nicolas Busca, Helion du Mas des Bourboux et al",
     author_email="nbusca@lpnhe.in2p3.fr",
-    packages=['picca','picca.fitter2'],
+    packages=['picca','picca.fitter2','picca_bin'],
     package_dir = {'': 'py'},
     package_data = {'picca': ['fitter2/models/*/*.fits']},
     install_requires=['numpy','scipy','iminuit','healpy','fitsio',
